MyConvenienceStore/src/test/java/org/example/Services/ServicesImpl/CashierServicesImplTest.java
package org.example.Services.ServicesImpl;

import org.example.Function.Fifo;
import org.example.StoreComponent.Cashier;
import org.example.StoreComponent.Customer;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.*;

public class CashierServicesImplTest {
    private CashierServicesImpl cashierServices;
    private Customer customer;
    private Cashier cashier;
    private CustomerServicesImpl customerService;

    @Before
        public void setUp() {
            // Initialize the CashierServicesImpl instance and Fifo queue
            cashierServices = new CashierServicesImpl();
            Fifo fifo = new Fifo(); // Replace with your Fifo queue implementation
        }


        @Test
        public void testDispensesReceiptsNoCustomerInQueue() {
            // Create a test scenario with no customer in the FIFO queue

            boolean result = cashierServices.dispensesReceipts(null); // Pass a valid or null cashier

            // Assert that the method returns false when no customer is in the queue
            assertFalse(result);
            // Add more assertions based on the expected behavior of the method
        }

        @Test
        public void testDispensesReceiptsNullCustomer() {
            // Create a test scenario with a null customer in the FIFO queue

            // Assuming you have a method to enqueue a null customer in the Fifo queue
            // Enqueue a null customer

            boolean result = cashierServices.dispensesReceipts(null); // Pass a valid or null cashier

            // Assert that the method returns false when the customer is null
            assertFalse(result);
            // Add more assertions based on the expected behavior of the method
        }

        // Add more test cases to cover different scenarios and edge cases

    }